Add files via upload
I did a simple regression using the gold values ​​of the items already in the game, but always using the items that have purely that factor, without mixing an item's ability or another factor for example: iron branch, chainmail, boot speed, attack claw .... Item's ability (dagon, vyse, lotus, true impact) would have an additional cost.

I had the idea when I wanted to compare the price of two talents on the same level in the talent tree, such as the level 10 slark: You can choose 6 of agi or 9 of strength, the cost of 6 of agi is already known (Band of Elvenskin) 450 gold but 9 strength would cost 863 gold. Following the same line of thought, I could compare a movespeed talent with damage and so on.

It can be useful for patch analysis, some comparison like that of talent or just out of curiosity.